Added CollapsedCardViewClass
